code optimisation

+++ app/src/main/java/com/example/mazzers/voicerecorder/utils/Utils.java
@@ -12,26 +12,27 @@ public class Utils {
      * Function to convert milliseconds time to
      * Timer Format
      * Hours:Minutes:Seconds
-     * */
-    public String milliSecondsToTimer(long milliseconds){
+     */
+    public String milliSecondsToTimer(long milliseconds) {
         //Log.d(TAG_LOG,"Utils: milliSecondsToTimer ");
         String finalTimerString = "";
         String secondsString = "";
 
         // Convert total duration into time
-        int hours = (int)( milliseconds / (1000*60*60));
-        int minutes = (int)(milliseconds % (1000*60*60)) / (1000*60);
-        int seconds = (int) ((milliseconds % (1000*60*60)) % (1000*60) / 1000);
+        int hours = (int) (milliseconds / (1000 * 60 * 60));
+        int minutes = (int) (milliseconds % (1000 * 60 * 60)) / (1000 * 60);
+        int seconds = (int) ((milliseconds % (1000 * 60 * 60)) % (1000 * 60) / 1000);
         // Add hours if there
-        if(hours > 0){
+        if (hours > 0) {
             finalTimerString = hours + ":";
         }
 
         // Prepending 0 to seconds if it is one digit
-        if(seconds < 10){
+        if (seconds < 10) {
             secondsString = "0" + seconds;
-        }else{
-            secondsString = "" + seconds;}
+        } else {
+            secondsString = "" + seconds;
+        }
 
         finalTimerString = finalTimerString + minutes + ":" + secondsString;
 
@@ -41,10 +42,11 @@ public class Utils {
 
     /**
      * Function to get Progress percentage
+     *
      * @param currentDuration
      * @param totalDuration
-     * */
-    public int getProgressPercentage(long currentDuration, long totalDuration){
+     */
+    public int getProgressPercentage(long currentDuration, long totalDuration) {
         //Log.d(TAG_LOG,"Utils: get progressPercentage");
         Double percentage = (double) 0;
 
@@ -52,7 +54,7 @@ public class Utils {
         long totalSeconds = (int) (totalDuration / 1000);
 
         // calculating percentage
-        percentage =(((double)currentSeconds)/totalSeconds)*100;
+        percentage = (((double) currentSeconds) / totalSeconds) * 100;
 
         // return percentage
         return percentage.intValue();
@@ -60,28 +62,28 @@ public class Utils {
 
     /**
      * Function to change progress to timer
-     * @param progress -
-     * @param totalDuration
-     * returns current duration in milliseconds
-     * */
+     *
+     * @param progress      -
+     * @param totalDuration returns current duration in milliseconds
+     */
     public int progressToTimer(int progress, int totalDuration) {
         //Log.d(TAG_LOG,"Utils: progressToTimer");
         int currentDuration = 0;
-        totalDuration = (int) (totalDuration / 1000);
-        currentDuration = (int) ((((double)progress) / 100) * totalDuration);
+        totalDuration = (totalDuration / 1000);
+        currentDuration = (int) ((((double) progress) / 100) * totalDuration);
 
         // return current duration in milliseconds
         return currentDuration * 1000;
     }
 
-    public String timeToString (int i){
+    public String timeToString(int i) {
         String time;
         int hours = i / 3600;
         int remainder = i - hours * 3600;
         int mins = remainder / 60;
         remainder = remainder - mins * 60;
         int secs = remainder;
-        time = hours+":"+mins+":"+secs+"";
+        time = hours + ":" + mins + ":" + secs + "";
         return time;
     }
 }
